move App component into app package
