+++ src/index.tsx
@@ -1,6 +1,6 @@
-import * as React from "react"
-import * as ReactDOM from "react-dom";
-import { App } from 'app';
+import * as React from 'react'
+import * as ReactDOM from 'react-dom';
+import { App } from "app/App";
 import "./style.css";
 
 const initialCardStacks = App.initializeStacksFromLocalStorage();
